chatbot/requirements.txt
numpy
streamlit
google-generativeai
python-dotenv
Pillow
